examples/image-process-sam/face_detection/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
pillow
facenet-pytorch